refactor(server): moved ErrorResponse from pkg/errors to unexported errorResponse in pkg/server
