+++ pkg/server/helpers_test.go
@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/shinzonetwork/shinzo-indexer-client/pkg/constants"
-	authErrors "github.com/shinzonetwork/shinzo-indexer-client/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -30,8 +29,29 @@ func TestWriteJSONError_ResponseBody(t *testing.T) {
 	t.Parallel()
 	rec := httptest.NewRecorder()
 	writeJSONError(rec, http.StatusUnauthorized, "unauthorized", "missing or empty credentials")
-	var resp authErrors.ErrorResponse
+	var resp errorResponse
 	require.NoError(t, json.Unmarshal(rec.Body.Bytes(), &resp))
 	assert.Equal(t, "unauthorized", resp.Code)
 	assert.Equal(t, "missing or empty credentials", resp.Message)
+}
+
+func TestErrorResponse_JSONSerialization(t *testing.T) {
+	t.Parallel()
+	resp := errorResponse{Code: "unauthorized", Message: "missing or empty credentials"}
+	data, err := json.Marshal(resp)
+	if err != nil {
+		t.Fatalf("json.Marshal() error: %v", err)
+	}
+
+	var decoded map[string]string
+	if err := json.Unmarshal(data, &decoded); err != nil {
+		t.Fatalf("json.Unmarshal() error: %v", err)
+	}
+
+	if decoded["error"] != "unauthorized" {
+		t.Errorf("json[error] = %q, want %q", decoded["error"], "unauthorized")
+	}
+	if decoded["message"] != "missing or empty credentials" {
+		t.Errorf("json[message] = %q, want %q", decoded["message"], "missing or empty credentials")
+	}
 }
